+++ src/main/php/de/thekid/dialog/App.php
@@ -11,18 +11,33 @@ use web\frontend\helpers\{Assets, Dates, Numbers};
 use web\frontend\{Frontend, AssetsFrom, AssetsManifest, HandlersIn, Handlebars};
 use web\rest\{RestApi, ResourcesIn};
 
+/** @test de.thekid.dialog.unittest.ServeTest */
 class App extends Application {
+  private $conn= null;
+  private $storage= null;
+
+  /** Passes MongoDB connection to use */
+  public function connecting(MongoConnection $conn): self {
+    $this->conn= $conn;
+    return $this;
+  }
+
+  /** Passes storage to use */
+  public function serving(Storage $storage): self {
+    $this->storage= $storage;
+    return $this;
+  }
 
   /** Returns routing for this web application */
   public function routes() {
     $preferences= new Preferences($this->environment, 'config');
-    $conn= new MongoConnection($preferences->get('mongo', 'uri'));
-    $repository= new Repository($conn->database($preferences->optional('mongo', 'db', 'dialog')));
-    $storage= new Storage($this->environment->arguments()[0]);
+    $this->conn??= new MongoConnection($preferences->get('mongo', 'uri'));
+    $this->storage??= new Storage($this->environment->arguments()[0]);
+    $repository= new Repository($this->conn->database($preferences->optional('mongo', 'db', 'dialog')));
     $inject= new Injector(Bindings::using()
       ->instance($repository)
-      ->instance($storage)
-      ->instance(new Signing(new Secret($preferences->get('mongo', 'uri')))->tolerating(seconds: 30))
+      ->instance($this->storage)
+      ->instance(new Signing(new Secret($this->conn->protocol()->dsn(password: true)))->tolerating(seconds: 30))
     );
 
     // Authenticate API users against MongoDB
@@ -34,17 +49,21 @@ class App extends Application {
     }
 
     // Cache static content for one week, immutable fingerprinted assets for one year
+    $caching= ['Cache-Control' => 'max-age=604800'];
     $manifest= new AssetsManifest($this->environment->path('src/main/webapp/assets/manifest.json'));
-    $static= ['Cache-Control' => 'max-age=604800'];
+    $static= new AssetsFrom($this->environment->path('src/main/webapp'))->with($caching);
+    $assets= new AssetsFrom($this->environment->path('src/main/webapp'))->with(fn($file) => [
+      'Cache-Control' => $manifest->immutable($file) ?? 'max-age=604800, must-revalidate'
+    ]);
+
     return [
-      '/image'      => $storage->with($static),
-      '/static'     => new AssetsFrom($this->environment->path('src/main/webapp'))->with($static),
-      '/assets'     => new AssetsFrom($this->environment->path('src/main/webapp'))->with(fn($file) => [
-        'Cache-Control' => $manifest->immutable($file) ?? 'max-age=604800, must-revalidate'
-      ]),
-      '/robots.txt' => fn($req, $res) => $res->send("User-agent: *\nDisallow: /api/\n", 'text/plain'),
-      '/api'        => $auth->optional(new RestApi(new ResourcesIn('de.thekid.dialog.api', $inject->get(...)))),
-      '/'           => new Frontend(
+      '/static'      => $static,
+      '/favicon.ico' => $static,
+      '/robots.txt'  => $static,
+      '/assets'      => $assets,
+      '/image'       => $this->storage->with($caching),
+      '/api'         => $auth->optional(new RestApi(new ResourcesIn('de.thekid.dialog.api', $inject->get(...)))),
+      '/'            => new Frontend(
         new HandlersIn('de.thekid.dialog.web', $inject->get(...)),
         new Handlebars($this->environment->path('src/main/handlebars'), [
           new Dates(TimeZone::getByName('Europe/Berlin')),