+++ src/main/php/de/thekid/dialog/api/Entries.php
@@ -71,7 +71,7 @@ class Entries {
           'meta'     => (array)$meta + ['dateTime' => gmdate('c')],
           'is'       => [$is => true]
         ];
-        foreach ($images ?? [] as $i => $existing) {
+        foreach ($images as $i => $existing) {
           if ($name === $existing['name']) {
             $images[$i]= $image;
             goto set;
@@ -105,7 +105,18 @@ class Entries {
 
   #[Put('/{id:.+(/.+)?}/published')]
   public function publish(#[Value] $user, string $id, #[Entity] Date $date) {
-    $this->repository->modify($id, ['$set' => ['published' => $date]]);
+
+    // If this entry does not contain any images, use the first image of the latest
+    // child element as the preview image. This will update the preview image every
+    // time new content is added.
+    $entry= $this->repository->entry($id, published: false);
+    if (empty($entry['images'])) {
+      $latest= $this->repository->children($id)->first();
+      $preview= empty($latest['images']) ? null : ['slug' => $latest['slug'], ...$latest['images'][0]];
+    } else {
+      $preview= ['slug' => $id, ...$entry['images'][0]];
+    }
+    $this->repository->modify($id, ['$set' => ['published' => $date, 'preview' => $preview]]);
 
     return ['published' => $date];
   }